stream: use ConcurrentHashMap to track in-progress async callbacks
Refs #29607

In the existing code, adding a callback and cleaning up callbacks contend
for updating the AtomicReference.

The way it is implemented, adding a callback invocation would invalidate an ongoing
cleanup run. The cleanup run is running directly from the GraphInterpreter,
so that in the worst case, cleanup would never be finished and the next loop
would be even slower because it would have to run through even more entries.

As long as new callbacks arrive, the stream would be completely stalled running
the cleanup code in a never ending loop.

The new code tries to decouple adding and cleaning up entries using a
ConcurrentHashMap. This will inherit any contention issues from CHM
but intuitively it should fare much better than the previous version:
The contention is between multiple threads adding entries to the map and
one thread (GraphInterpreter) removing entries. The keys
(identityHashCodes of promises) are all independent of each other, so
assuming that collision means potential contention, the table should
perform well enough if it is sized adequately. We currently hope that
the default size and settings is good enough.